Give the Ex command index a type of its own
`cmdidx_T` was `int` and the 560 command indexes were `pub const CMD_*:
CMD_index`, so `exarg_T.cmdidx` compared equal to any integer in the
program and c2rust's `switch (cmdidx)` arms survived as
`match idx as c_int { c if c == CMD_catch as c_int => … }`. The table is
apigen's already -- `types/cmdidx.rs` and `ex_docmd/cmdtable.rs` come out
of the same pass over `ex_cmds.lua` -- so the enum is emitted, not
written.

`CmdIdx` is `#[repr(i32)]` with every discriminant spelled out, because a
built-in's discriminant *is* its row in `cmdnames` and `cmdidxs1`/
`cmdidxs2` are precomputed offsets into that same order. `USER`/
`USER_BUF` keep -1/-2 and `SIZE` keeps the count, which doubles as the
"not a command" marker.

**A variant is spelled the way the command is** -- `append`, `Next`,
`r#if`. Capitalising to Rust's convention collides twelve pairs that
differ only in case (`:next` and `:Next` are different commands), and
twelve more command names are Rust keywords. `#[allow(non_camel_case_types)]`
buys back an identifier that is the thing it names.

Five `mem::zeroed::<exarg_T>()` sites and `exarg`'s `Default` rely on
all-zero meaning `:append`, which the explicit `append = 0` keeps true.

Where the sweep could not be mechanical:

- `find_ex_command`'s scan walks `usize` rows and names what it stopped at
  once (`CmdIdx::at_row`), rather than stepping an enum. So does
  `excmd_get_cmdidx`.
- `cmdidx < 0` is `is_user_cmd`, which is now a `matches!` over the two
  negative variants rather than a sign test.
- `skip_cmd`'s 78-command table became a `matches!` alternation, and the
  four `c if c == …` guard arms in `do_one_cmd`/`ex_operators`/
  `get_user_command_name` became real patterns.

The `-O0` bill, which is what both test suites build, is why the last two
matter: a guard arm mid-`match` stops rustc building a switch, and a
linear `.contains` over enum values is one `PartialEq` call per element.
For the same reason `PartialEq` is hand-written and `#[inline(always)]`
-- the derived `eq` is an ordinary call, and roughly a thousand sites
compare a `cmdidx` against a command.

`one_letter_cmd` answers `bool` rather than 0/1, paying for the
`-> c_int` that `CmdIdx::code` adds. `types/name_uniqueness.rs`'s
"the parser still sees declarations" floor drops from 1,000 to 900:
560 `pub const`s became one enum.

`const_int_alias` 3045 -> 2485 · `pub_items` 12,854 -> 12,293 ·
`t_suffix_types` 216 -> 215 · `files_without_deny_casts` 1051 -> 1050 ·
`unsafe_lines_outside_perimeter` 54,724 -> 54,722. No layout moved:
`cmdidx_T` was `int` and `CmdIdx` is `#[repr(i32)]`.

+++ crates/nvim/src/quickfix/stack.rs
@@ -166,7 +166,7 @@ pub(crate) fn qf_cmd_stack(eap: Ea, print_emsg: bool) -> Option<Qi> {
 /// quickfix command works on the global one and answers `None`.
 pub(crate) fn qf_cmd_stack_or_alloc(eap: Ea) -> (Qi, Option<Win>) {
     // SAFETY: a command's `cmdidx` is one of the table's.
-    if !unsafe { is_loclist_cmd(eap.cmdidx as c_int) } {
+    if !unsafe { is_loclist_cmd(eap.cmdidx) } {
         return (QfStack::Global.qi(), None);
     }
     let wp = cur_win();
@@ -625,7 +625,7 @@ pub(crate) unsafe fn qf_cmd_get_stack(eap: *mut exarg_T, print_emsg: bool) -> *m
     // SAFETY: the caller's promise -- a live `exarg_T`.
     let eap = unsafe { Ea::new(eap) };
     // SAFETY: forwarded from the caller.
-    if !unsafe { is_loclist_cmd(eap.cmdidx as c_int) } {
+    if !unsafe { is_loclist_cmd(eap.cmdidx) } {
         return QfStack::Global.raw();
     }
     let qi = win_loclist(cur_win());
